+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+
+Fix:
+  - Remove warning error from unknown `Button` props.
+
 ## [2.9.1] - 2019-06-04
 
 Fix: